# apps/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Health Watchers API",
    "version": "1.0.0",
    "description": "HIPAA-compliant healthcare management platform API with Stellar blockchain payment integration."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "hw_ API key",
        "description": "Clinic API key (`hw_\u2026`) with SMART-style scopes: `patient/*.read` or `patient/<Type>.read`. `Authorization: ApiKey hw_\u2026` is also accepted."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "FhirOperationOutcome": {
        "type": "object",
        "required": ["resourceType", "issue"],
        "properties": {
          "resourceType": {
            "type": "string",
            "enum": ["OperationOutcome"]
          },
          "issue": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "severity": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "diagnostics": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "FhirBundle": {
        "type": "object",
        "required": ["resourceType", "type"],
        "properties": {
          "resourceType": {
            "type": "string",
            "enum": ["Bundle"]
          },
          "type": {
            "type": "string",
            "enum": ["searchset"]
          },
          "total": {
            "type": "integer"
          },
          "link": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "relation": {
                  "type": "string"
                },
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "entry": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "fullUrl": {
                  "type": "string"
                },
                "resource": {
                  "type": "object"
                },
                "search": {
                  "type": "object",
                  "properties": {
                    "mode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/Patient/{id}": {
      "servers": [
        {
          "url": "/fhir/r4",
          "description": "FHIR R4 read API"
        }
      ],
      "get": {
        "summary": "Read a FHIR Patient",
        "tags": ["FHIR R4"],
        "security": [
          {
            "apiKeyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resourceType": {
                      "type": "string",
                      "enum": ["Patient"]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found in the API key's clinic",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "403": {
            "description": "API key lacks the patient/<Type>.read (or patient/*.read) scope",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "429": {
            "description": "API key rate limit exceeded",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          }
        }
      }
    },
    "/Patient": {
      "servers": [
        {
          "url": "/fhir/r4",
          "description": "FHIR R4 read API"
        }
      ],
      "get": {
        "summary": "Search FHIR Patients",
        "tags": ["FHIR R4"],
        "security": [
          {
            "apiKeyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "query",
            "description": "`[system|]value`; system is https://healthwatchers.com/patient-id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_count",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "_cursor",
            "in": "query",
            "description": "Opaque — follow Bundle.link[relation=next] instead of building it.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "searchset Bundle with self/next paging links",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirBundle"
                }
              }
            }
          },
          "400": {
            "description": "Invalid search parameter",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "403": {
            "description": "API key lacks the patient/<Type>.read (or patient/*.read) scope",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "429": {
            "description": "API key rate limit exceeded",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          }
        }
      }
    },
    "/Encounter": {
      "servers": [
        {
          "url": "/fhir/r4",
          "description": "FHIR R4 read API"
        }
      ],
      "get": {
        "summary": "Search FHIR Encounters for a patient",
        "tags": ["FHIR R4"],
        "security": [
          {
            "apiKeyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient",
            "in": "query",
            "required": true,
            "description": "`Patient/<id>` or `<id>`",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_count",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "_cursor",
            "in": "query",
            "description": "Opaque — follow Bundle.link[relation=next] instead of building it.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "searchset Bundle with self/next paging links",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirBundle"
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid patient parameter",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "403": {
            "description": "API key lacks the patient/<Type>.read (or patient/*.read) scope",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "429": {
            "description": "API key rate limit exceeded",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          }
        }
      }
    },
    "/Observation": {
      "servers": [
        {
          "url": "/fhir/r4",
          "description": "FHIR R4 read API"
        }
      ],
      "get": {
        "summary": "Search vital-sign Observations for a patient",
        "tags": ["FHIR R4"],
        "security": [
          {
            "apiKeyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "patient",
            "in": "query",
            "required": true,
            "description": "`Patient/<id>` or `<id>`",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "`vital-signs` (only category available)",
            "schema": {
              "type": "string",
              "example": "vital-signs"
            }
          },
          {
            "name": "_count",
            "in": "query",
            "description": "Encounters per page; each contributes its vital signs",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "_cursor",
            "in": "query",
            "description": "Opaque — follow Bundle.link[relation=next] instead of building it.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "searchset Bundle with self/next paging links",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirBundle"
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid patient parameter",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "403": {
            "description": "API key lacks the patient/<Type>.read (or patient/*.read) scope",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          },
          "429": {
            "description": "API key rate limit exceeded",
            "content": {
              "application/fhir+json": {
                "schema": {
                  "$ref": "#/components/schemas/FhirOperationOutcome"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}